requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
redis
requests
beautifulsoup4
playwright
python-dotenv
tqdm
openai
